tests/wpt/web-platform-tests/css/css-contain/content-visibility/content-visibility-026.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Content Visibility: Computed Values</title>
<link rel="author" title="Rakina Zata Amni" href="mailto:rakina@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert" content="content-visibility:hidden does not affect computed value of 'contain'">
<meta name="assert" content="content-visibility:auto does not affect computed value of 'contain'">

<div id="container"></div>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>

test(() => {
  assert_equals(getComputedStyle(container).contain, "none");
  container.style = "content-visibility:hidden";
  assert_equals(getComputedStyle(container).contentVisibility, "hidden");
  assert_equals(getComputedStyle(container).contain, "none");
}, "content-visibility:hidden does not affect computed value of 'contain'");

test(() => {
  assert_equals(getComputedStyle(container).contain, "none");
  container.style = "content-visibility:auto;contain:layout;";
  assert_equals(getComputedStyle(container).contentVisibility, "auto");
  assert_equals(getComputedStyle(container).contain, "layout");
}, "content-visibility:auto does not affect computed value of 'contain'");

</script>
